+++ src/main/java/shadows/apotheosis/ench/EnchModule.java
@@ -87,7 +87,7 @@ import shadows.apotheosis.ench.library.EnchLibraryBlock;
 import shadows.apotheosis.ench.library.EnchLibraryContainer;
 import shadows.apotheosis.ench.library.EnchLibraryTile.BasicLibraryTile;
 import shadows.apotheosis.ench.library.EnchLibraryTile.EnderLibraryTile;
-import shadows.apotheosis.ench.objects.GlowyItem;
+import shadows.apotheosis.ench.objects.GlowyBlockItem;
 import shadows.apotheosis.ench.objects.ScrappingTomeItem;
 import shadows.apotheosis.ench.objects.TomeItem;
 import shadows.apotheosis.ench.replacements.BaneEnchant;
@@ -295,11 +295,11 @@ public class EnchModule {
 				new TomeItem(Items.BOW, EnchantmentCategory.BOW).setRegistryName("bow_tome"),
 				new ScrappingTomeItem(),
 				new BlockItem(Apoth.Blocks.HELLSHELF, new Item.Properties().tab(Apotheosis.APOTH_GROUP)).setRegistryName("hellshelf"),
-				new GlowyItem(Apoth.Blocks.INFUSED_HELLSHELF, new Item.Properties().tab(Apotheosis.APOTH_GROUP)).setRegistryName("infused_hellshelf"),
+				new GlowyBlockItem(Apoth.Blocks.INFUSED_HELLSHELF, new Item.Properties().tab(Apotheosis.APOTH_GROUP)).setRegistryName("infused_hellshelf"),
 				new BlockItem(Apoth.Blocks.BLAZING_HELLSHELF, new Item.Properties().tab(Apotheosis.APOTH_GROUP)).setRegistryName("blazing_hellshelf"),
 				new BlockItem(Apoth.Blocks.GLOWING_HELLSHELF, new Item.Properties().tab(Apotheosis.APOTH_GROUP)).setRegistryName("glowing_hellshelf"),
 				new BlockItem(Apoth.Blocks.SEASHELF, new Item.Properties().tab(Apotheosis.APOTH_GROUP)).setRegistryName("seashelf"),
-				new GlowyItem(Apoth.Blocks.INFUSED_SEASHELF, new Item.Properties().tab(Apotheosis.APOTH_GROUP)).setRegistryName("infused_seashelf"),
+				new GlowyBlockItem(Apoth.Blocks.INFUSED_SEASHELF, new Item.Properties().tab(Apotheosis.APOTH_GROUP)).setRegistryName("infused_seashelf"),
 				new BlockItem(Apoth.Blocks.CRYSTAL_SEASHELF, new Item.Properties().tab(Apotheosis.APOTH_GROUP)).setRegistryName("crystal_seashelf"),
 				new BlockItem(Apoth.Blocks.HEART_SEASHELF, new Item.Properties().tab(Apotheosis.APOTH_GROUP)).setRegistryName("heart_seashelf"),
 				new BlockItem(Apoth.Blocks.ENDSHELF, new Item.Properties().tab(Apotheosis.APOTH_GROUP)).setRegistryName("endshelf"),